+++ Kernel/dlTBox.h
@@ -663,7 +663,7 @@ protected:	// methods
 
 		/// init Extra Rule field in concepts given by a vector V with a given INDEX
 	inline void
-	initRuleFields ( const ConceptVector& v, unsigned int index ) const
+	initRuleFields ( const ConceptVector& v, size_t index ) const
 	{
 		for ( ConceptVector::const_iterator q = v.begin(), q_end = v.end(); q < q_end; ++q )
 			(*q)->addExtraRule(index);
@@ -1134,7 +1134,7 @@ public:
 		/// GCI Axioms access
 	BipolarPointer getTG ( void ) const { return T_G; }
 		/// get simple rule by its INDEX
-	const TSimpleRule* getSimpleRule ( unsigned int index ) const { return SimpleRules[index]; }
+	const TSimpleRule* getSimpleRule ( size_t index ) const { return SimpleRules[index]; }
 
 		/// check if the relevant part of KB contains inverse roles.
 	bool isIRinQuery ( void ) const
